Submit acceleration structure rebuild
Some refactoring around how queues are handled - we want queues to
have a reference to Device, but since queues are owned by the device,
this gets tricky. In order to avoid circular Arc references or a cascade
of lifetimes, construct the actual Queue struct as needed. These are no
more than a vulkan handle and some metadata, and don't require any
destruction so we're free to clone em and pass em around as much as
needed.